+++ sage-android/Sage/app/src/main/java/blueprint/com/sage/main/fragments/AdminPanelFragment.java
@@ -14,16 +14,16 @@ import java.util.HashMap;
 import java.util.List;
 
 import blueprint.com.sage.R;
-import blueprint.com.sage.browse.BrowseMentorsActivity;
-import blueprint.com.sage.browse.BrowseSchoolsActivity;
+import blueprint.com.sage.admin.browse.BrowseMentorsActivity;
+import blueprint.com.sage.admin.browse.BrowseSchoolsActivity;
 import blueprint.com.sage.events.semesters.SemesterListEvent;
 import blueprint.com.sage.models.Semester;
 import blueprint.com.sage.network.Requests;
-import blueprint.com.sage.requests.CheckInRequestsActivity;
-import blueprint.com.sage.requests.SignUpRequestsActivity;
-import blueprint.com.sage.semester.CreateSemesterActivity;
-import blueprint.com.sage.semester.FinishSemesterActivity;
-import blueprint.com.sage.semester.SemesterListActivity;
+import blueprint.com.sage.admin.requests.CheckInRequestsActivity;
+import blueprint.com.sage.admin.requests.SignUpRequestsActivity;
+import blueprint.com.sage.admin.semester.CreateSemesterActivity;
+import blueprint.com.sage.admin.semester.FinishSemesterActivity;
+import blueprint.com.sage.admin.semester.SemesterListActivity;
 import blueprint.com.sage.shared.interfaces.BaseInterface;
 import blueprint.com.sage.utility.network.NetworkUtils;
 import blueprint.com.sage.utility.view.FragUtils;
