+++ sage-android/Sage/app/src/main/java/blueprint/com/sage/main/fragments/UserFragment.java
@@ -23,7 +23,7 @@ import blueprint.com.sage.shared.interfaces.BaseInterface;
 import blueprint.com.sage.shared.interfaces.ListDialogInterface;
 import blueprint.com.sage.shared.interfaces.ToolbarInterface;
 import blueprint.com.sage.shared.views.CircleImageView;
-import blueprint.com.sage.users.EditUserActivity;
+import blueprint.com.sage.users.profile.EditUserActivity;
 import blueprint.com.sage.utility.network.NetworkUtils;
 import blueprint.com.sage.utility.view.FragUtils;
 import butterknife.Bind;
